lab2/consumer/requirements.txt
kafka-python==2.0.2
cassandra-driver==3.29.1
ujson==5.10.0
